requirements.txt
Django==1.9.5
django-categories==1.4.1
django-mptt==0.8.3
django-taggit==0.18.0
mysqlclient==1.3.7
Pillow==3.2.0
requests==2.9.1
unicode-slugify==0.1.1
